+++ packages/headless/src/controllers/core/facets/_common/facet-option-definitions.ts
@@ -38,3 +38,19 @@ const facetSearchOptionDefinitions: SchemaDefinition<FacetSearchRequestOptions> 
 export const facetSearch = new RecordValue({
   values: facetSearchOptionDefinitions,
 });
+
+export const allowedValues = new RecordValue({
+  options: {required: false},
+  values: {
+    type: new StringValue({
+      constrainTo: ['simple'],
+      emptyAllowed: false,
+      required: true,
+    }),
+    values: new ArrayValue({
+      required: true,
+      max: 25,
+      each: new StringValue({emptyAllowed: false, required: true}),
+    }),
+  },
+});