+++ packages/headless/src/controllers/facets/facet/headless-facet.test.ts
@@ -57,6 +57,7 @@ describe('facet', () => {
       field: 'author',
       sortCriteria: 'score',
       facetSearch: {},
+      allowedValues: ['foo', 'bar'],
     };
 
     state = createMockState();
@@ -87,10 +88,10 @@ describe('facet', () => {
 
     initFacet();
 
-    expect(FacetIdDeterminor.determineFacetId).toHaveBeenCalledWith(
-      engine,
-      options
-    );
+    expect(FacetIdDeterminor.determineFacetId).toHaveBeenCalledWith(engine, {
+      ...options,
+      allowedValues: {type: 'simple', values: ['foo', 'bar']},
+    });
   });
 
   it('registers a facet with the passed options and the default values of unspecified options', () => {
@@ -101,6 +102,7 @@ describe('facet', () => {
       filterFacetCount: true,
       injectionDepth: 1000,
       numberOfValues: 8,
+      allowedValues: {type: 'simple', values: ['foo', 'bar']},
     });
 
     expect(engine.actions).toContainEqual(action);